Revert "removed position rounding in library"
This reverts commit 88e3db574a8bbcfede6eb059179238f8f4900b70.

Conflicts:
	gwt-connectors/src/main/java/pl/tecna/gwt/connectors/client/ConnectionPoint.java
	gwt-connectors/src/main/java/pl/tecna/gwt/connectors/client/Diagram.java
	gwt-connectors/src/main/java/pl/tecna/gwt/connectors/client/drag/EndPointDragController.java
	gwt-connectors/src/main/java/pl/tecna/gwt/connectors/client/elements/EndPoint.java
	gwt-connectors/src/main/java/pl/tecna/gwt/connectors/client/elements/Shape.java

+++ gwt-connectors/src/main/java/pl/tecna/gwt/connectors/client/CornerPoint.java
@@ -2,7 +2,7 @@ package pl.tecna.gwt.connectors.client;
 
 public class CornerPoint extends Point {
 
-  public CornerPoint(double left, double top) {
+  public CornerPoint(Integer left, Integer top) {
     super(left, top);
   }
